% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmFS.r
\name{glmFS_step2}
\alias{glmFS_step2}
\title{glmFS Step 2}
\usage{
glmFS_step2(step1_output_dir, EN_cutoff = 50, anova_baseline = NA)
}
\arguments{
\item{step1_output_dir}{WIP}

\item{EN_cutoff}{WIP}

\item{anova_baseline}{WIP}
}
\value{
WIP
}
\description{
Run the second step of glmFS
}
\examples{
WIP
}
