% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACIFIC.r
\name{PACIFIC}
\alias{PACIFIC}
\title{Run PACIFIC feature selection pipeline}
\usage{
PACIFIC(
  data,
  baseline = c("age", "sex", "stage"),
  feat1,
  feat2,
  discretization_method = "median",
  sparsity_threshold = "5_percent",
  univariate_p_cutoff = 0.1,
  subsampling_ratio = 0.8,
  num_iterations = 10,
  EN_cutoff = 50,
  output_dir,
  verbose = TRUE
)
}
\arguments{
\item{data}{data.frame. The table of input data with samples in rows and features in columns.}

\item{baseline}{character vector. The baseline clinical features as columns of "data". Default: c("age", "sex", "stage", "grade"). Put NA for empty set.}

\item{feat1}{character vector. The feature set 1 as columns of "data".}

\item{feat2}{character vector. The feature set 2 as columns of "data". Default: NA (for empty set)}

\item{discretization_method}{character. Continuous features are binarized using this method within the feature selection loops. Default: "median"}

\item{sparsity_threshold}{character. Sparse features are called using this threshold within the feature selection loops. Default: "5_percent"}

\item{univariate_p_cutoff}{numeric. Features weakly associated with survival are called using this cutoff for P-value of univariate CoxPH models within the feature selection loops. Default: 0.1}

\item{subsampling_ratio}{numeric. Ratio used for subsampling the "data" for the feature selection loops. Default: 0.8}

\item{num_iterations}{numeric. How many iterations of the feature selection loop to run. Default: 10}

\item{output_dir}{character. The directory to store the results.}

\item{verbose}{character. Whether to print progress messages. Default: FALSE}
}
\description{
Run PACIFIC feature selection pipeline
}
