% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACIFIC.r
\name{PACIFIC_step1}
\alias{PACIFIC_step1}
\title{PACIFIC Step 1}
\usage{
PACIFIC_step1(
  data,
  response,
  event = NA,
  single_features = NA,
  interaction_features = NA,
  features_to_discretize = NA,
  discretization_method = "median",
  features_with_flexible_direction = NA,
  features_to_skip_sparsity_prefiltering = NA,
  features_to_skip_univariate_association_prefiltering = NA,
  sparsity_criterion = "5_percent",
  univariate_p_cutoff = 0.1,
  subsampling_ratio = 0.8,
  num_iterations = 10,
  output_dir,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame (or an extension of data.framethe, e.g. data.table). The table of input data with rows for samples and columns for features and variables.}

\item{response}{Name of the column in data which contains the response values.}

\item{event}{Name of the column in 'data' which contains the event status values for survival outcomes. This column must contain only 0 and 1. Default is 'NA'.}

\item{single_features}{(for input features) Vector of "single features". Default is 'NA' (for no single feature).}

\item{interaction_features}{(for input features) Explicit or list-based specification of "interaction features". Explicit: a vector of the form 'c("A*B", "C*D", ...)'. List-based: list of two vectors of single components for all two-way combinations as interaction features. Default is 'NA' (for no interaction feature). Note that 'single_features' cannot overlap with any component in 'interaction_features'.}

\item{features_to_discretize}{The subset of input features (being single or component of interaction) to be discretized, i.e. to be converted from numeric to categorical (subject to 'discretization_method'). Default is 'NA' (for no such feature).}

\item{discretization_method}{The method for discretization. Either "median" (for median dichotomization), or a 'function' designed to get a numeric vector as input and return a factor vector (of the same length) as output. Default is '"median"'.}

\item{features_with_flexible_direction}{The subset of input features (being single or component of interaction) to be allowed to have flexible direction. Applicable only to either factor variables, or numeric variables in 'features_to_discretize'. This means that the reference level of these variables are to be determined by the pipeline in order to optimize the associations with the 'response' Default is 'NA' (for no such feature).}

\item{features_to_skip_sparsity_prefiltering}{The subset of input features (being single only) to be skipped for prefiltering by sparsity (within iterations). Default is 'NA' (for no such feature).}

\item{features_to_skip_univariate_association_prefiltering}{The subset of input features (being single only) to be skipped for prefiltering by weak associations with the 'response' (within iterations) . Default is NA (for no such feature).}

\item{sparsity_criterion}{The criterion used for detecting sparse features in prefiltering by sparsity (within iterations). Applicable to features which involve factor(s). It should be a string of the form '"X_percent"' or '"X_absolute"' (where 'X' is a number), meaning that the cutoff for the minimum size of the feature is either 'X' percent of the size of the input data (within the iteration) or is just 'X', respectively. Default is '"5_percent"'.}

\item{univariate_p_cutoff}{Cutoff for the P values of univariate models used for prefiltering by weak association with the 'response'. Default is 0.1.}

\item{subsampling_ratio}{The ratio used for subsampling 'data' to make the input data for each iteration. Note that the ratio of events (in survival analysis) or each level of response (if it's categorical) in 'data' is maintained in the subsamplings. Default is '0.8'.}

\item{num_iterations}{Number of iterations to run. Default is '10'.}

\item{output_dir}{The directory to store the results of iterations.}

\item{verbose}{Whether to print progress messages. Default is 'FALSE'.}
}
\description{
Run the first step of PACIFIC
}
