% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACIFIC.r
\name{PACIFIC_step2}
\alias{PACIFIC_step2}
\title{PACIFIC Step 2}
\usage{
PACIFIC_step2(
  step1_output_dir,
  EN_cutoff = 50,
  anova_baseline = NA,
  verbose = FALSE
)
}
\arguments{
\item{step1_output_dir}{WIP}

\item{EN_cutoff}{WIP}

\item{anova_baseline}{WIP}
}
\value{
WIP
}
\description{
Run the second step of PACIFIC
}
\examples{
WIP
}
