% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACIFIC.r
\name{PACIFIC_survival_step2}
\alias{PACIFIC_survival_step2}
\title{Run step 2 of PACIFIC for survival outcomes}
\usage{
PACIFIC_survival_step2(
  output_dir,
  EN_cutoff = 50,
  anova_baseline = NA,
  verbose = TRUE
)
}
\description{
Run step 2 of PACIFIC for survival outcomes
}
