% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvFS.r
\name{SurvFS_step1}
\alias{SurvFS_step1}
\title{SurvFS Step 1}
\usage{
SurvFS_step1(
  data,
  time = NA,
  status = NA,
  outcome = NA,
  single_features = NA,
  interaction_features = NA,
  features_to_discretize = NA,
  discretization_method = "median",
  features_with_flexible_direction = NA,
  features_to_skip_sparsity_prefiltering = NA,
  features_to_skip_univariate_association_prefiltering = NA,
  sparsity_criterion = "5_percent",
  univariate_p_cutoff = 0.1,
  subsampling_ratio = 0.8,
  num_iterations = 10,
  output_dir,
  verbose = FALSE
)
}
\arguments{
\item{data}{WIP}

\item{time}{WIP}

\item{status}{WIP}

\item{outcome}{WIP}

\item{single_features}{WIP}

\item{interaction_features}{WIP}

\item{features_to_discretize}{WIP}

\item{discretization_method}{WIP}

\item{features_with_flexible_direction}{WIP}

\item{features_to_skip_sparsity_prefiltering}{WIP}

\item{features_to_skip_univariate_association_prefiltering}{WIP}

\item{sparsity_criterion}{WIP}

\item{univariate_p_cutoff}{WIP}

\item{subsampling_ratio}{WIP}

\item{num_iterations}{WIP}

\item{output_dir}{WIP}

\item{verbose}{WIP}
}
\value{
WIP
}
\description{
Run the first step of SurvFS
}
\examples{
WIP
}
