% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvFS.r
\name{current_total_iters}
\alias{current_total_iters}
\title{Current Total Iterations}
\usage{
current_total_iters(dir)
}
\arguments{
\item{dir}{WIP}
}
\value{
WIP
}
\description{
Get the total number of iterations currently stored in a directory
}
\examples{
WIP
}
