% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvFS.r
\name{SurvFS_step1}
\alias{SurvFS_step1}
\title{SurvFS Step 1}
\usage{
SurvFS_step1(
  data,
  time = "time",
  status = "status",
  single_features = NA,
  interaction_features = NA,
  features_to_discretize = NA,
  discretization_method = "median",
  SurvFS_step2features_with_flexible_direction = NA,
  SurvFS_step2features_to_skip_sparsity_prefiltering = NA,
  SurvFS_step2features_to_skip_survival_prefiltering = NA,
  SurvFS_step2sparsity_criterion = "5_percent",
  SurvFS_step2wald_p_cutoff = 0.1,
  SurvFS_step2subsampling_ratio = 0.8,
  SurvFS_step2num_iterations = 10,
  SurvFS_step2output_dir,
  SurvFS_step2verbose = FALSE
)
}
\arguments{
\item{data}{WIP}

\item{time}{WIP}

\item{status}{WIP}

\item{single_features}{WIP}

\item{interaction_features}{WIP}

\item{features_to_discretize}{WIP}

\item{discretization_method}{WIP}

\item{features_with_flexible_direction}{WIP}

\item{features_to_skip_sparsity_prefiltering}{WIP}

\item{features_to_skip_survival_prefiltering}{WIP}

\item{sparsity_criterion}{WIP}

\item{wald_p_cutoff}{WIP}

\item{subsampling_ratio}{WIP}

\item{num_iterations}{WIP}

\item{output_dir}{WIP}

\item{verbose}{WIP}
}
\value{
WIP
}
\description{
Run the first step of SurvFS
}
\examples{
WIP
}
